If you supply -help as an option the translator tells you what arguments it supports.
The emake.bat batch files uses truncated names in the DOS as
names are truncated when opened under DOS.  Help is given if
you enter the parameters different than anticipated.
You can specify the compiler -gcc for UNIX systems.

Bind is fixed to call the backend as backendd.exe instead of
by its old name backend.exe when running in DOS mode
and if it is not found in EUDIR/bin then it will
check also the same directory as the ex file.

tests included: bindtest is an interactive to verify this works.

--HG--
extra : convert_revision : svn%3A52ce87f7-99e8-7440-b109-e810650da9f1/trunk%401320