Enhanced DEBUG_OPCODE_TRACE mode (enabled debugging compile time-C define) to also show parameters passed to the opcodes.
Before, the library if specified was relative to %EUDIR%\bin in some situations.  Now, if specified it is relative to the current directory of the translator's process.  If unspecified it runs relative to the directory where the translator resides. The translator now creates emake.bat and objfiles.lnk files that allow you to see the C code in the watcom debugger.  This doesn't work with most files in DOS mode, however.

--HG--
extra : convert_revision : svn%3A52ce87f7-99e8-7440-b109-e810650da9f1/trunk%401430